Refactored store utilities into `/src/utils/store.js`
