+++ src/utils.js
@@ -118,52 +118,6 @@ export function areSetsEqual(setA, setB) {
 }
 
 /**
- * jwtPayload(token)
- *
- * Returns the parsed payload object from the given JWT payload (does not attempt to validate it!
- * Don't trust the data you get out!)
- *
- * Source: https://stackoverflow.com/a/38552302/38666
- */
-export function jwtPayload(token) {
-  const base64Url = token.split('.')[1]
-  const base64 = base64Url.replace(/-/g, '+').replace(/_/g, '/')
-  const jsonPayload = decodeURIComponent(atob(base64).split('').map(function(c) {
-      return '%' + ('00' + c.charCodeAt(0).toString(16)).slice(-2)
-  }).join(''))
-
-  return JSON.parse(jsonPayload)
-}
-
-/**
- * localStoreFactory(rootKey)
- *
- * Factory funcion for generating local store access functions keyed off rootKey (that is, the
- * local store will include a single rootKey that contains a JSON object with whatever is set
- * via the factory-derived methods).
- */
-export function localStoreFactory(rootKey) {
-  function storeGetAll () {
-    const stored = window.localStorage.getItem(rootKey)
-    return stored ? JSON.parse(stored) : {}
-  }
-  function storeGet (key) {
-    const stored = storeGetAll()
-    return stored[key]
-  }
-  function storeSet (key, value) {
-    const stored = storeGetAll()
-    stored[key] = value
-    window.localStorage.setItem(rootKey, JSON.stringify(stored))
-  }
-  return {
-    storeGetAll,
-    storeGet,
-    storeSet,
-  }
-}
-
-/**
  * Returns phoenixborn image url from the CDN.
  *
  * @param {str} stub Phoenixborn card name
